Define ID ranges to use non-root containers
When using non-root container theres a limit on the number of subuids
and subids available to the container and to be able to login as an IPA
user (e.g. su admin) and to setup a trust the start and size of the
ranges must respect this limits.

This setup allowed the creation of the trust running as a regular user.