Use Github Action to run the idm-to-idm trust test
This patch uses Github action FreeIPA-Cluster-Test
(https://github.com/rjeffman/FreeIPA-Cluster-Test) to run the test
playbook using the clusted defined with the cluster configuration file.

The advantage of using the action is that it requires less code and it
works in the case that 'become: true' is used in the playboks (allowing
playbooks to be used also with VMs, that need privilege escalation).

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>